Create README.md
Safety is the most important aspect of any industry. Security and safety are absolutely essential in the mining industry. The mining industry follows a few fundamental safety measures to prevent any accidents. Underground mines still have accidents because of rising temperatures, rising water levels, and methane gas leaks. Here, we provide worker safety. He can call security by pressing the panic button when a worker is in danger. A reliable communication system between workers in underground mines and the fixed ground mine system is required to improve safety in underground mines. There must be no interruptions to the communication network at any time or under any circumstance. In this project, a low-cost, early-warning intelligence-based wireless mine supervision system is proposed. IOT allows for monitoring of worker status.Mines are the world’s most dangerous place to work because in the mines, explosion often happens and thousand people are dying.   
And a recent report states that in such mine accidents an average of around 12,000 people have died.   
Coal is a non-sustainable origin that cannot be widely replaced by humans, there are several mishaps of coalmines occurring in the mines.   
The diggers are putting their lives at risk, by working in the    coal mines, even once in a while they end up losing their lives in the coal mines that are an unfortunate part.